% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleAge.R
\docType{data}
\name{AvAgeModel_ls}
\alias{AvAgeModel_ls}
\title{AV age model}
\format{
A list having the following components:
\describe{
  \item{Auto}{a vector of cumulative probability of AV autos by age}
  \item{LtTrk}{a vector of cumulative probability of AV light trucks by age}
}
}
\source{
AssignVehicleAge.R script.
}
\usage{
AvAgeModel_ls
}
\description{
A list containing the AV age model probability tables
needed to implement the AV age model.
}
\keyword{datasets}
