% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVAVOrientation.R
\name{CVAVOrientation}
\alias{CVAVOrientation}
\title{Assign level of autonomous to household vehicles}
\usage{
CVAVOrientation(L)
}
\arguments{
\item{L}{A list containing data defined by the module specification.}
}
\value{
A list containing data produced by the function consistent with the
module specifications.
}
\description{
\code{CVAVOrientation} assigns level of automation to household vehicles
}
