% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVAVOrientation.R
\docType{data}
\name{CVAVOrientationSpecifications}
\alias{CVAVOrientationSpecifications}
\title{Specifications list for CVAVOrientation module}
\format{
A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}
}
\source{
CVAVOrientation.R script.
}
\usage{
CVAVOrientationSpecifications
}
\description{
A list containing specifications for the CVAVOrientation module.
}
\keyword{datasets}
